Migrate Fixed Method to use Data structure
