+++ src/fixed/mod.rs
@@ -1,4 +1,4 @@
 mod chars;
-mod database;
 pub(crate) mod method;
 mod parser;
+mod search;
